# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aamras — Premium Mango Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  slug: string;
  name: string;
  description: string;
  emoji?: string;
}

export interface Product {
  id: string;
  slug: string;
  name: string;
  variety: string;
  tagline: string;
  description: string;
  price: number;
  /** e.g. "per box (3kg)" */
  unit: string;
  imageUrl: string;
  gallery?: string[];
  origin: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  sweetness: number;
  rating: number;
  reviewCount: number;
  stock: number;
  deliveryEtaHours: number;
  badge?: string | null;
  categorySlug: string;
  isBestSeller: boolean;
  isOrganic: boolean;
}

export interface Review {
  id: string;
  productId?: string | null;
  authorName: string;
  authorLocation: string;
  avatarUrl: string;
  rating: number;
  title: string;
  body: string;
  createdAt: string;
}

export interface CartItem {
  id: string;
  productId: string;
  product: Product;
  quantity: number;
  lineTotal: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  itemCount: number;
}

export interface AddCartItemBody {
  sessionId: string;
  productId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface UpdateCartItemBody {
  sessionId: string;
  /** @minimum 0 */
  quantity: number;
}

export interface OrderItem {
  id: string;
  productId: string;
  productName: string;
  productImageUrl: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
}

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  email: string;
  phone: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  state: string;
  pincode: string;
  /** e.g. "upi", "card", "cod" */
  paymentMethod: string;
  items: OrderItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  status: string;
  estimatedDelivery: string;
  createdAt: string;
}

export interface CreateOrderBody {
  sessionId: string;
  customerName: string;
  email: string;
  phone: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  state: string;
  pincode: string;
  paymentMethod: string;
}

export interface SubscriptionPlan {
  id: string;
  slug: string;
  name: string;
  tagline: string;
  description: string;
  weeklyPrice: number;
  boxesPerWeek: number;
  varieties: string[];
  perks: string[];
  imageUrl: string;
}

export interface Subscription {
  id: string;
  planId: string;
  customerName: string;
  email: string;
  phone: string;
  startDate: string;
  status: string;
  createdAt: string;
}

export interface CreateSubscriptionBody {
  planId: string;
  customerName: string;
  email: string;
  phone: string;
}

export interface StatsSummary {
  varietiesCount: number;
  ordersToday: number;
  happyCustomers: number;
  averageDeliveryHours: number;
  farmsPartnered: number;
}

export type ListProductsParams = {
  /**
   * Category slug filter
   */
  category?: string;
  search?: string;
};

export type ListReviewsParams = {
  productId?: string;
};

export type GetCartParams = {
  sessionId: string;
};

export type RemoveCartItemParams = {
  sessionId: string;
};