# lib/api-zod/src/generated/types/addCartItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aamras — Premium Mango Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface AddCartItemBody {
  sessionId: string;
  productId: string;
  /** @minimum 1 */
  quantity: number;
}