# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aamras — Premium Mango Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface CartItem {
  id: string;
  productId: string;
  product: Product;
  quantity: number;
  lineTotal: number;
}