# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aamras — Premium Mango Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  sessionId: string;
  customerName: string;
  email: string;
  phone: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  state: string;
  pincode: string;
  paymentMethod: string;
}